chore: split util functions into their own files
