+++ src/core/utils/cookie.spec.ts
@@ -1,0 +1,15 @@
+import { validateCookie } from "./cookie";
+
+describe("validateCookie()", () => {
+  it("cookies = ''", () => {
+    expect(validateCookie("")).toBe(false);
+  });
+
+  it("cookies = 'abc'", () => {
+    expect(validateCookie("")).toBe(false);
+  });
+
+  it("cookies = 'foo=bar'", () => {
+    expect(validateCookie("foo=bar")).toBe(false);
+  });
+});
